requirements/local.txt
Django==1.9.2
Fabric==1.11.1
djangorestframework==3.3.3
django-oauth-toolkit==0.10.0
django-extensions==1.7.3
django-cors-headers==1.1.0
rest-framework-generic-relations==1.1.0
drf-nested-routers==0.11.1
Markdown==2.6.6
uuid==1.30
